[Zephyr][platform] Rework the architecture of the DiagnosticDataProviderImpl.h/cpp
* Keep DiagnosticDataProviderImpl.h/cpp base class generic
* Add a subclass specific to Wi-Fi network diagnostics, to avoid multiple conditions (#ifdef) in the base class file
* This change improves readability, separation of concerns and better maintainability
