+++ src/platform/Zephyr/DiagnosticDataProviderImpl.h
@@ -53,23 +53,6 @@ public:
     CHIP_ERROR GetNetworkInterfaces(NetworkInterface ** netifpp) override;
     void ReleaseNetworkInterfaces(NetworkInterface * netifp) override;
 
-#ifdef CONFIG_WIFI
-    CHIP_ERROR GetWiFiBssId(MutableByteSpan & value) override;
-    CHIP_ERROR GetWiFiSecurityType(app::Clusters::WiFiNetworkDiagnostics::SecurityTypeEnum & securityType) override;
-    CHIP_ERROR GetWiFiVersion(app::Clusters::WiFiNetworkDiagnostics::WiFiVersionEnum & wiFiVersion) override;
-    CHIP_ERROR GetWiFiChannelNumber(uint16_t & channelNumber) override;
-    CHIP_ERROR GetWiFiRssi(int8_t & rssi) override;
-    CHIP_ERROR GetWiFiBeaconLostCount(uint32_t & beaconLostCount) override;
-    CHIP_ERROR GetWiFiBeaconRxCount(uint32_t & beaconRxCount) override;
-    CHIP_ERROR GetWiFiPacketMulticastRxCount(uint32_t & packetMulticastRxCount) override;
-    CHIP_ERROR GetWiFiPacketMulticastTxCount(uint32_t & packetMulticastTxCount) override;
-    CHIP_ERROR GetWiFiPacketUnicastRxCount(uint32_t & packetUnicastRxCount) override;
-    CHIP_ERROR GetWiFiPacketUnicastTxCount(uint32_t & packetUnicastTxCount) override;
-    CHIP_ERROR GetWiFiCurrentMaxRate(uint64_t & currentMaxRate) override;
-    CHIP_ERROR GetWiFiOverrunCount(uint64_t & overrunCount) override;
-    CHIP_ERROR ResetWiFiNetworkDiagnosticsCounts() override;
-#endif // CONFIG_WIFI
-
 protected:
     DiagnosticDataProviderImpl();
